jira-cost-impact-bot/requirements.txt
slack_sdk
requests
python-docx
